service-backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.3
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.2
